+++ sismic/model.py
@@ -23,9 +23,9 @@ class Event:
         self.data = additional_parameters
 
     def __eq__(self, other):
-        return isinstance(other, Event) and \
-               self.name == other.name and \
-               self.data == other.data
+        return (isinstance(other, Event) and
+                self.name == other.name and
+                self.data == other.data)
 
     def __getattr__(self, attr):
         try:
@@ -847,6 +847,13 @@ class MicroStep:
     def __repr__(self):
         return 'MicroStep({}, {}, >{}, <{})'.format(self.event, self.transition, self.entered_states, self.exited_states)
 
+    def __eq__(self, other):
+        return (isinstance(other, MicroStep) and
+                self.event == other.event and
+                self.transition == other.transition and
+                self.entered_states == other.entered_states and
+                self.exited_states == other.exited_states)
+
 
 class MacroStep:
     """
@@ -914,3 +921,8 @@ class MacroStep:
     def __repr__(self):
         return 'MacroStep@{}({}, {}, >{}, <{})'.format(round(self.time, 3), self.event, self.transitions,
                                                        self.entered_states, self.exited_states)
+
+    def __eq__(self, other):
+        return (isinstance(other, MacroStep) and
+                self.time == other.time and
+                self.steps == other.steps)