+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ The YAML format of a statechart also changed, look carefully at the changelog an
   ``transitions_with``, ``remove_transition`` and ``rotate_transition``.
 - (Added) Statechart: new methods to manipulate states: ``remove_state``, ``rename_state``, ``move_state``,
   ``state_for``, ``parent_for``, ``children_for``.
+- (Added) Steps: add ``__eq__`` on ``MacroStep`` and ``MicroStep``.
 - (Added) Module: a new exceptions hierarchy (see ``exceptions`` module).
   The new exceptions are used in place of the old ones (``Warning``, ``AssertionError`` and ``ValueError``).
 - (Changed) YAML: uppermost *states:* should be replaced by *initial state:* and can contain at most one state.